+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes in git
 --------------
 
 * Added ``delete_translation()`` API.
+* Fix creating translations which only have default values.
 
 
 Changes in 1.6.5 (2016-07-11)